AMR: Bug fix in AMReX library
E-mail from Andrew Myers:

Okay, I think I see what's going on. We are using a form of interpolation
in AssignDensityFort that doesn't necessarily conserve global minima.
Could you try the following change to AMReX_ParticleContainerI.H:

https://github.com/AMReX-Codes/amrex/commit/e3af48538baa9e9cf3fa70d6176ca66d7bf528c1

(or alternatively, you could switch to the development version of AMReX) and see if that fixes the problem?